+++ src/Payment/Contracts/Payment.php
@@ -33,13 +33,13 @@ interface Payment
 
     public function getMethod(): PaymentMethod;
 
-    /**
-     * @todo add these in v4
-     * public function getSubtype(): ?string
-     * public function hasRemoteId(): bool
-     * public function getRemoteId(): ?string
-     * public function isOffline(): bool
-     */
+    public function getSubtype(): ?string;
+
+    public function hasRemoteId(): bool;
+
+    public function getRemoteId(): ?string;
+
+    public function isOffline(): bool;
 
     public function getPayable(): Payable;