+++ Changelog.md
@@ -101,6 +101,11 @@
 - BC: Added the `getConfigurationSchema()` method to the `Configurable` interface
 - BC: Changed the `PaymentMethod` interface into Configurable
 - BC: Added the `getRemoteId()` method to the `PaymentRequest` interface
+- BC: Added the following methods to the `Payment` interface:
+  - `getSubtype()`
+  - `hasRemoteId()`
+  - `getRemoteId()`
+  - `isOffline()`
 - BC: Added the following methods to the `Payable` interface:
     - `getNumber()`
     - `getPayableRemoteId()`